refactor deepseek v3 2 directory and remove get_input_sharding
